feat (UI): Users can delete and edit activity values
