+++ app/src/i18n/locales/en/data.json
@@ -777,6 +777,7 @@
   "add-emission-data-btn": "Add emission data",
   "update-activity": "Update activity",
   "add-emission-data": "Add emission data",
+  "update-emission-data": "Update emission data",
   "co2-emissions": "CO2 emissions",
   "n2o-emissions": "N2O emissions",
   "ch4-emissions": "CH4 emissions",
